src/mob/share/requirements.txt
opencv-python
RPi.GPIO
pyserial
smbus
simple-pid
matplotlib
numpy
